Uses separate project types file
