+++ api/project/project.types.ts
@@ -1,0 +1,20 @@
+export enum ProjectLanguage {
+  react = "react",
+  node = "node",
+  python = "python",
+}
+
+export enum ProjectStatus {
+  info = "info",
+  warning = "warning",
+  error = "error",
+}
+
+export type Project = {
+  id: string;
+  name: string;
+  language: ProjectLanguage;
+  numIssues: number;
+  numEvents24h: number;
+  status: ProjectStatus;
+};
